Changes in technicians js
